heroku-requirements.txt
gunicorn==19.9.0
django_heroku==0.3.1
whitenoise==4.1.1
